requirements.txt
fastapi
uvicorn[standard]
pydantic
pymysql
python-dotenv
pandas
xgboost
joblib
sqlalchemy
scikit-learn
requests
sentence-transformers
qdrant-client
